ai_worker/requirements.txt
mysql-connector-python==8.3.0
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
joblib==1.3.2
